requirements.txt
# Core ETL Libraries
pandas==2.1.4
pydantic==2.5.3
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Airflow
apache-airflow==2.8.3

# Configuration
pyyaml==6.0.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.4
pytest-cov==4.1.0

# Data validation
email-validator==2.1.0

# Additional utilities
openpyxl==3.10.10
